backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
Pillow>=10.0.0
rembg>=2.0.50
onnxruntime>=1.16.0
trimesh>=4.0.0
numpy>=1.24.0
pygltflib==1.16.3
slowapi>=0.1.8  # Rate limiting for API
httpx>=0.27.0
matplotlib>=3.8.0
